reset password page security tweak
It is commonly accepted security best practice on a modern system that a reset password page should be non-commital about if a user existed or not. This is to prevent simple attacks against the reset password page to determine if a user is registered on a system or not by email address or username. Such attacks could then be combined with other attacks to break into user accounts.

As such i have changed the way the reset page returns its status to always return success. This would need to be combined with further tweaks to change the wording, i am not capable of making this change as it would involve changes to I18N content.

Please view this PR as a start of a longer discussion as to how to handle this simple tweak.