`authorize` method may be removed if not used
both removing the `authorize()` method, and returning `true` will result in the same behavior, so adding a comment to explain this is an option.

https://github.com/laravel/framework/blob/10.x/src/Illuminate/Foundation/Http/FormRequest.php#L190

I would even go so far as to say we should *recommend* removing the method, as it's less code to run through, but this is probably good enough to start.