[11.x] add clarification on component slot escaping (#9813)
* add clarification on component slot escaping

this was a little confusing to me when I first saw it, so wanted to add some clarification.

it is likely, especially when using a component for your layout, that the `$slot` variable will contain HTML.  however, the docs show the `{{ $slot }}` escaped version. although I'm not exactly sure where this happens in the code, it seems that a slot is treated specially, and does not escape content.  All code inside the slot appears to respect any escaped or unescaped content.

I've also added a wrapping `<div>` around each task to more closely reflect a real world use, and to further drive home the `$slot` will handle any HTML content it is provided.

* Update blade.md

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>