[11.x] Add support for union types in event listeners in Laravel 11 (#9957)
* [11.x] Add support for union types in event listeners in Laravel 11

It's something I simply discovered. It simplifies some of my listeners, so I thought it would be helpful for other developers to know that this is possible. By using union types in Laravel event listeners, you can handle multiple related events in a single listener. This reduces redundancy and makes it easier to manage similar events in one place.

The example in the documentation shows how to use union types in the `handle` method, leveraging the `instanceof` operator to differentiate and handle each event accordingly.

* Update events.md

* Update events.md

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>