use the word "Blog" instead of "Post"
the word "post" can have multiple contexts in the world of web development, so in order to impove searchability, let's avoid using "Post" as a Model/Noun.

this will allow people searching for things related to the HTTP verb "POST" and route related things in these docs a little easier, without losing any clarity in the rest of the page.