Update mail.md (#10151)
* Update mail.md

To send email with Mailgun (and any external mail service, probably) and Laravel, there are three changes to config files that have to be made.  Two of them had code examples in the documentation.  The third one did not.  Currently, it is just written with words as "change the default value..." but the fact that the item in the array is named "default" and value in the key/value pair is "default" leads to ambiguity.

After reading Laraval mail's documentation, doing a lot of set up and configuring, and programming a controller and mailable, I experienced hours of stress and frustration because the emails weren't being sent.  All of the Mailgun env and config vars seemed to be good.  This was the problem.

Please accept this mail documentation change.  The one-line code example is important in getting mail to work

* Update mail.md

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>