remove arbitrary chaining alignment
use a single indent for chains on a new line, rather than trying to align it with some arbitrary pattern in the first line. this arbitrary alignment very easily gets out of sync as changes occur to the code, and often times the adjustments are not made to the chained alignment. the single indent alignment does not suffer this problem.

this is to bring us in line with https://github.com/laravel/framework/pull/53835 and https://github.com/laravel/framework/pull/53748

this commit only contains whitespace changes.

if accepted, I will make these adjustments to the remaining docs pages.