remove `<?php` opening tag in code blocks
we're being explicit now with our fenced code blocks, rather than relying on guessing, so we don't need to give the guesser any extra help.

this also brings consistency because we have a mix of code blocks that use it and ones that don't.

also, IMO, it's just noise on the example