standardize ellipsis usage (#10187)
in order to indicate "more content" or "content not relevant to example", we often use an ellipsis. in PHP and other languages, this pattern actually is relevant to the language.

this change standardizes usage of the ellipsis for the above examples to be in a comment. this give us valid PHP and Javascipt, and reduced ambiguity with the language construct.