[12.x] Rework pagination Tailwind JIT description; add v4 syntax (#10238)
* chore: rework pagination Tailwind JIT description; add Tailwind v4 syntax

In Tailwind CSS v3, JIT mode is enabled by default (which was introduced in v2). I believe that Laravel 12.x is far enough from Tailwind CSS v2 that we can drop this naming convention and simply refer to it as v3.

* https://v3.tailwindcss.com/docs/upgrade-guide#migrating-to-the-jit-engine

The release of Tailwind CSS v4 introduced many breaking changes. Automatic source detection was introduced, but unfortunately, the vendor folder was excluded due to the .gitignore file. Adding the source now works with a different syntax compared to v3.

* https://tailwindcss.com/docs/detecting-classes-in-source-files#which-files-are-scanned
* https://tailwindcss.com/docs/functions-and-directives#source-directive

* wip: use " instead of '

* Update pagination.md

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>