[12.x] clarify how to "set" value objects with custom casts (#10342)
* clarify how to "set" value objects with custom casts

the current wording only accounts for 1 of the 2 scenarios that can occur when using a custom cast with a value object. that Value Object can encompasses either 1 or many values on the model.

the current documentation addresses the example of an `Address`, which would have multiple fields like "line1", "line2", "city", etc.  this `set` method should return the array as currently documented.

an example of a single value Value Object would be a phone number. a user might want it as an value object rather than a raw number to encapsulate behavior. this `set` method should return a string.

there is one distinct advantage to returning a string vs an array. it prevents the cast from being tied to explicit model field names. this allows a field of any name to use the cast:

```php
'phone' => AsPhone::class,
'telephone' => AsPhone::class,
```

if a user returns an array from this simple 1 value cast, the cast loses its flexibility and the above example does not work.

* Update eloquent-mutators.md

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>