Update responses.md to make easier for new users to use stream package. (#10466)
* Update responses.md to make easier for new users to use stream package.

Totally understand if these changes don't get merged. Spent a day trying to understand the new streaming docs and had to go to the source of the package to see why I was having issues.

The Laravel stream package is hardcoded to use POST requests yet the second example is a GET request. Updating to POST to make it easier to onboard new developers.

The CSRF token is not alway present in the head of a layout. Make sure to call this out otherwise new users see 419 without a good error message on how to address.

* Update responses.md

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>