only use and document the `limit` and `offset` methods
`skip()` and `take()` are alias methods to `offset()` and `limit()`. IMO we should only use and document the main methods, not the aliases.  This will encourage more performant coding, just like we now do in `laravel/framework`.

IF it's decided we still want to mention the aliases, I'd recommended listing them second as the alternatives.  I can make that change if desired.