Eloquent - Changed wording on Deleting models
The wording was such that it seemed like the `destroy` method would not load the models (which would be a huge performance benefit). However, this is not the case! The benefit here is that _you_ don't have to load it first, before deletion.

So I changed the wording to make it less ambiguous and move the warning to a more appropriate place.