document default slot usage
while we always have the option for `$slot` content to live anywhere in the component, we can also be explicit by using the unnamed `<x-slot>` tag. this is necessary when you want to pass attributes to your default slot.

this is kind of documentation for https://github.com/laravel/framework/pull/56883 but the ability to have a default slot actually already existed, you just needed to name it "slot".

`<x-slot:slot>`