clarify `where()`, `whereNull()`, and `whereNotNull()` usage (#10814)
this change helps clarify usage of the `where()`, `whereNull()`, and `whereNotNull()` methods with a `null` value.

discourage the use of `$collection->where('field', '!=', null)` because it will possibly cause unexpected results with "nully" values.

show more values that are "nully" and how they will be treated with the strict check.