[12.x] explain how to use `@once`/`@pushOnce` with a custom ID (#10867)
* explain how to use `@once`/`@pushOnce` with a custom ID

relying on the default random generated string only works if the code is compiled once, like in a loop or in a single component.

if you wish to use the directives across multiple compilations, like in separate components, you **must** pass a custom unique ID for them to only be rendered once.

* formatting

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>