update documentation for custom collections
there are 3 ways to define a custom collection, 2 of which are currently documented

- `CollectedBy` attribute
- override `newCollection()` method
- override `$collectionClass` property (undocumented)

surprisingly, the undocumented option is arguably the best. it doesn't have the performance hit from Reflection of the attribute, and doesn't force the complete override of a parent method which may miss out on upstream changes.

I went aggressive in this commit, and removed the 2 current options in favor of the single best one. However, I could see the argument to keep the "attribute" option, and would add that back in if desired.